backend_v2/requirements/base.txt
# Core Django and DRF
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0

# Filtering support for DRF
django-filter>=23.5,<24.0

# CORS handling for React frontend
django-cors-headers>=4.3,<5.0

# Image processing for model ImageFields
Pillow>=12.0.0 

# Environment variable management
python-decouple>=3.8,<4.0

djangorestframework-simplejwt>=5.3,<6.0

whitenoise

bleach>=6.0,<7.0

# === SEO & Analysis Dependencies (NEW) ===
beautifulsoup4>=4.12,<5.0          # bs4 for HTML parsing in analysis
textstat>=0.7,<1.0                 # Readability scores (Flesch etc.)
extruct>=0.18,<0.19                # Schema.org / metadata extraction
django-ratelimit>=4.1,<5.0         # Rate limiting in API views

# Optional but recommended (your apps.py and settings reference them)
django-meta>=2.0,<3.0              # django-meta for OG/Twitter tags
django-robots>=6.0,<7.0            # django-robots for robots.txt

# Celery stack (used heavily in tasks.py)
celery>=5.3,<6.0
django-celery-beat>=2.9,<3.0       # Periodic tasks
redis>=5.0,<6.0                    # Broker (you already reference Redis in settings)


# Optional: Postgres driver if you switch from SQLite later
# psycopg2-binary>=2.9,<3.0
django-admin-numeric-filter>=0.1,<0.2